restructured files removed useless code commented out shader logic
